feat: add @Pipe and more class
