+++ packages/core/src/decorator/common/objectDef.ts
@@ -1,6 +1,6 @@
 import { saveObjectDefinition } from '../decoratorManager';
-import { ScopeEnum } from '../interface';
 import { Provide } from './provide';
+import { ScopeEnum } from '../../interface';
 
 export function Init(): MethodDecorator {
   return function (target: any, propertyKey: string) {
