+++ packages/core/src/decorator/common/pipeline.ts
@@ -1,6 +1,6 @@
 import { createCustomPropertyDecorator } from '../decoratorManager';
 import { PIPELINE_IDENTIFIER } from '../constant';
-import { ObjectIdentifier } from '../interface';
+import { ObjectIdentifier } from '../../interface';
 
 export function Pipeline(
   valves?: Array<ObjectIdentifier | (new (...args) => any)>
