+++ packages/core/src/decorator/common/provide.ts
@@ -1,5 +1,5 @@
 import { saveProviderId } from '../decoratorManager';
-import { ObjectIdentifier } from '../interface';
+import { ObjectIdentifier } from '../../interface';
 
 export function Provide(identifier?: ObjectIdentifier) {
   return function (target: any) {
