+++ packages/core/src/decorator/decoratorManager.ts
@@ -3,12 +3,14 @@ import {
   GroupModeType,
   IModuleStore,
   InjectModeEnum,
+  MethodDecoratorOptions,
   ObjectDefinitionOptions,
   ObjectIdentifier,
   ParamDecoratorOptions,
   TagClsMetadata,
   TagPropsMetadata,
-} from './interface';
+  TSDesignType,
+} from '../interface';
 import {
   INJECT_CUSTOM_METHOD,
   INJECT_CUSTOM_PARAM,
@@ -711,13 +713,7 @@ export function clearAllModule() {
   return manager.clear();
 }
 
-export interface TSDesignType {
-  name: string;
-  originDesign: any;
-  isBaseType: boolean;
-}
-
-function transformTypeFromTSDesign(designFn): TSDesignType {
+export function transformTypeFromTSDesign(designFn): TSDesignType {
   if (isNullOrUndefined(designFn)) {
     return { name: 'undefined', isBaseType: true, originDesign: designFn };
   }
@@ -968,8 +964,14 @@ export function createCustomPropertyDecorator(
 export function createCustomMethodDecorator(
   decoratorKey: string,
   metadata: any,
-  impl = true
+  implOrOptions: boolean | MethodDecoratorOptions = { impl: true }
 ): MethodDecorator {
+  if (typeof implOrOptions === 'boolean') {
+    implOrOptions = { impl: implOrOptions } as MethodDecoratorOptions;
+  }
+  if (implOrOptions.impl === undefined) {
+    implOrOptions.impl = true;
+  }
   return function (target: any, propertyName: string, descriptor) {
     attachClassMetadata(
       INJECT_CUSTOM_METHOD,
@@ -977,7 +979,7 @@ export function createCustomMethodDecorator(
         propertyName,
         key: decoratorKey,
         metadata,
-        impl,
+        options: implOrOptions,
       },
       target
     );
