+++ packages/core/src/decorator/web/requestMapping.ts
@@ -1,7 +1,8 @@
 /**
  * 'HEAD', 'OPTIONS', 'GET', 'PUT', 'PATCH', 'POST', 'DELETE' 封装
  */
-import { attachClassMetadata, WEB_ROUTER_KEY, MiddlewareParamArray } from '../';
+import { attachClassMetadata, WEB_ROUTER_KEY } from '../';
+import { MiddlewareParamArray } from '../../interface';
 
 export interface RouterOption {
   /**
