+++ packages/core/src/service/lifeCycleService.ts
@@ -1,4 +1,4 @@
-import { ILifeCycle, IMidwayContainer } from '../interface';
+import { ILifeCycle, IMidwayContainer, ScopeEnum } from '../interface';
 import {
   CONFIGURATION_KEY,
   Init,
@@ -6,7 +6,6 @@ import {
   listModule,
   Provide,
   Scope,
-  ScopeEnum,
 } from '../decorator';
 import { FunctionalConfiguration } from '../functional/configuration';
 import { MidwayFrameworkService } from './frameworkService';
