+++ packages/core/src/service/middlewareService.ts
@@ -1,10 +1,11 @@
-import { Provide, Scope, ScopeEnum } from '../decorator';
+import { Provide, Scope } from '../decorator';
 import {
   CommonMiddleware,
   IMiddleware,
   IMidwayContainer,
   FunctionMiddleware,
   IMidwayApplication,
+  ScopeEnum,
 } from '../interface';
 import { MidwayCommonError, MidwayParameterError } from '../error';
 import { isIncludeProperty, pathMatching } from '../util';
