+++ packages/core/src/service/mockService.ts
@@ -2,8 +2,9 @@ import {
   IMidwayApplication,
   IMidwayContainer,
   IMidwayContext,
+  ScopeEnum,
 } from '../interface';
-import { Destroy, Init, Provide, Scope, ScopeEnum } from '../decorator';
+import { Destroy, Init, Provide, Scope } from '../decorator';
 
 @Provide()
 @Scope(ScopeEnum.Singleton)
