+++ packages/core/src/service/pipelineService.ts
@@ -1,5 +1,5 @@
-import { IMidwayContainer } from '../interface';
-import { getProviderName, ObjectIdentifier } from '../decorator';
+import { IMidwayContainer, ObjectIdentifier } from '../interface';
+import { getProviderName } from '../decorator';
 
 interface IPipelineInfo {
   /**
