+++ packages/validate/src/constants.ts
@@ -1,2 +1,3 @@
 export const RULES_KEY = 'common:rules';
 export const VALIDATE_KEY = 'common:validate';
+export const VALID_KEY = 'validate:valid';
