+++ packages/validate/src/index.ts
@@ -1,6 +1,8 @@
 export { ValidateConfiguration as Configuration } from './configuration';
 export * from './decorator/rule';
 export * from './decorator/validate';
+export * from './decorator/valid';
 export * from './dtoHelper';
 export * from './service';
 export * from './error';
+export * from './pipe';
